Case Study/E-StockMarketApplication/src/main/java/com/example/EStockMarketApplication/Exceptions/GlobalExceptionHandler.java
package com.example.EStockMarketApplication.Exceptions;

import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.ControllerAdvice;
import org.springframework.web.bind.annotation.ExceptionHandler;

@ControllerAdvice
public class GlobalExceptionHandler {
    @ExceptionHandler(CompanyNotFound.class)
    public ResponseEntity<String> handleCompanyNotFoundException(CompanyNotFound e)
    {
        return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
    }

    @ExceptionHandler(LessTurnOverException.class)
    public ResponseEntity<String> handleLessTurnOverExceptionException(LessTurnOverException e)
    {
        return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
    }
}
